Add fordfulkersonkoegel algorithm
